# docs/public/openapi.yaml
openapi: 3.0.3
info:
  title: TeslaSync API
  version: 1.0.0
  description: Self-hosted Tesla Fleet Intelligence API
  license:
    name: MIT
  contact:
    name: TeslaSync
    url: https://github.com/ev-dev-labs/teslasync
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /vehicles/{vehicleID}:
    get:
      operationId: getVehicle
      summary: Get single vehicle
      description: Returns details for a specific vehicle.
      tags: [Vehicles]
      parameters:
        - $ref: "#/components/parameters/vehicleID"
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteVehicle
      summary: Delete vehicle and associated data
      description: Deletes a vehicle and all its associated data (drives, charging, positions, etc.).
      tags: [Vehicles]
      parameters:
        - $ref: "#/components/parameters/vehicleID"
      responses:
        "200":
          description: Vehicle deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: vehicle deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /drives/{driveID}:
    get:
      operationId: getDrive
      summary: Get single drive
      description: Returns details for a specific drive session.
      tags: [Drives]
      parameters:
        - $ref: "#/components/parameters/driveID"
      responses:
        "200":
          description: Drive details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drive"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    vehicleID:
      name: vehicleID
      in: path
      required: true
      description: Vehicle ID
      schema:
        type: integer
    driveID:
      name: driveID
      in: path
      required: true
      description: Drive session ID
      schema:
        type: integer
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: resource not found
            code: DB_RECORD_NOT_FOUND
            category: database
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: an internal error occurred
            code: INTERNAL_ERROR
            category: internal
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: resource not found
        code:
          type: string
          description: Machine-readable error code
          example: DB_RECORD_NOT_FOUND
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - UNPROCESSABLE_ENTITY
            - AUTH_INVALID_CREDENTIALS
            - AUTH_TOKEN_EXPIRED
            - AUTH_TOKEN_INVALID
            - AUTH_TOKEN_MISSING
            - AUTH_API_KEY_INVALID
            - AUTH_API_KEY_EXPIRED
            - AUTH_API_KEY_REVOKED
            - AUTH_INSUFFICIENT_PERMISSIONS
            - VEHICLE_NOT_FOUND
            - VEHICLE_OFFLINE
            - VEHICLE_ASLEEP
            - VEHICLE_BUSY
            - DRIVE_NOT_FOUND
            - CHARGE_NOT_FOUND
            - TRIP_NOT_FOUND
            - SESSION_STILL_ACTIVE
            - SESSION_NOT_FOUND
            - TESLA_API_UNAVAILABLE
            - TESLA_API_SUSPENDED
            - TESLA_API_RATE_LIMITED
            - TESLA_API_TIMEOUT
            - TESLA_API_AUTH_FAILED
            - TESLA_NOT_CONNECTED
            - DB_CONNECTION_FAILED
            - DB_QUERY_FAILED
            - DB_TRANSACTION_FAILED
            - DB_RECORD_NOT_FOUND
            - DB_DUPLICATE_RECORD
            - VALIDATION_INVALID_INPUT
            - VALIDATION_INVALID_JSON
            - VALIDATION_MISSING_FIELD
            - VALIDATION_INVALID_RANGE
            - VALIDATION_INVALID_ID
            - VALIDATION_PAYLOAD_TOO_LARGE
            - GEOFENCE_NOT_FOUND
            - GEOFENCE_INVALID_COORDINATES
            - GEOFENCE_INVALID_RADIUS
            - COMMAND_NOT_SUPPORTED
            - COMMAND_FAILED
            - COMMAND_TIMEOUT
            - BACKUP_CONFIG_NOT_FOUND
            - BACKUP_RUN_NOT_FOUND
            - BACKUP_FAILED
            - BACKUP_STORAGE_ERROR
            - RESTORE_FAILED
            - NOTIFICATION_CHANNEL_NOT_FOUND
            - NOTIFICATION_TEST_FAILED
            - TELEMETRY_INGEST_FAILED
            - MQTT_UNAVAILABLE
            - EXPORT_FAILED
            - EXPORT_NOT_FOUND
            - EXPORT_INVALID_FORMAT
            - RATE_LIMITED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - GATEWAY_TIMEOUT
            - NOT_IMPLEMENTED
            - ERROR
        category:
          type: string
          description: Error category
          example: database
          enum:
            - auth
            - vehicle
            - drives
            - charging
            - trips
            - tesla_api
            - database
            - validation
            - geofence
            - commands
            - backup
            - notifications
            - telemetry
            - export
            - rate_limiting
            - internal
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          description: Internal database ID
          example: 1
        vehicle_id:
          type: integer
          description: Tesla vehicle ID
          example: 1234567890
        vin:
          type: string
          description: Vehicle identification number
          example: 5YJ3E1EA1LF000000
        display_name:
          type: string
          description: User-assigned display name
          example: My Tesla
        model:
          type: string
          description: Vehicle model
          example: Model 3
        trim_badging:
          type: string
          description: Trim level
          example: Long Range
        exterior_color:
          type: string
          description: Exterior color
          example: Pearl White
        wheel_type:
          type: string
          description: Wheel type
          example: Aero
        state:
          type: string
          description: Current vehicle state
          example: online
        healthy:
          type: boolean
          description: Whether the vehicle data is healthy
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Drive:
      type: object
      description: |
        SI canonical drive aggregate (Phase-42 migration 000185, Phase-48
        Slice 1 rename). All distance/speed/energy/power values are in SI base
        units. Convert at the display boundary using `useUnits()` on the
        frontend.
      properties:
        id:
          type: integer
          example: 1
        vehicle_id:
          type: integer
          example: 1
        start_ts:
          type: string
          format: date-time
        end_ts:
          type: string
          format: date-time
          nullable: true
        start_lat:
          type: number
          example: 37.7749
        start_lon:
          type: number
          example: -122.4194
        end_lat:
          type: number
          example: 37.3382
        end_lon:
          type: number
          example: -121.8863
        start_address:
          type: string
          nullable: true
          example: San Francisco, CA
        end_address:
          type: string
          nullable: true
          example: San Jose, CA
        distance_m:
          type: number
          description: Distance travelled in meters (SI canonical).
          example: 124761.0
        duration_s:
          type: integer
          format: int64
          description: Drive duration in seconds (SI canonical).
          example: 3312
        avg_speed_mps:
          type: number
          description: Average speed in meters per second (SI canonical).
          nullable: true
          example: 23.4
        max_speed_mps:
          type: number
          description: Maximum speed in meters per second (SI canonical).
          nullable: true
          example: 33.5
        start_battery_pct:
          type: integer
          nullable: true
          example: 85
        end_battery_pct:
          type: integer
          nullable: true
          example: 62
        energy_used_wh:
          type: number
          description: Energy used in watt-hours (SI canonical).
          nullable: true
          example: 15300
        regen_energy_wh:
          type: number
          description: Energy recovered via regenerative braking in watt-hours (SI canonical).
          nullable: true
          example: 1850
        avg_power_w:
          type: number
          description: Average power draw in watts (SI canonical).
          nullable: true
          example: 16500
        outside_temp_avg_c:
          type: number
          description: Average ambient temperature in degrees Celsius.
          nullable: true
          example: 22.5
...